requirements.txt
torch==2.2.0
numpy==1.24.4
biopython==1.81
matplotlib==3.8.0
scikit-learn==1.3.0
tensorflow==2.15.0
